# openapi/openapi-payment.json
{
	"openapi": "3.1.0",
	"info": {
	  "title": "IWM Backend API",
	  "description": "Documentation for the modular monolith hexagonal backend APIs.",
	  "version": "0.1.0-SNAPSHOT"
	},
	"servers": [
	  {
		"url": "https://kernel-core.yowyob.com/kernel-api",
		"description": "Generated server url"
	  }
	],
	"security": [
	  {
		"ClientId": [],
		"ApiKey": []
	  },
	  {
		"ClientId": [],
		"ApiKey": [],
		"BearerAuth": []
	  }
	],
	"paths": {
...
	  "/api/commercial-plans/{planCode}/auto-renewal": {
		"post": {
		  "tags": [
			"plan-commercial-alias-controller"
		  ],
		  "operationId": "enableAutoRenewal",
		  "parameters": [
			{
			  "name": "planCode",
			  "in": "path",
			  "required": true,
			  "schema": {
				"type": "string"
			  }
			}
		  ],
		  "requestBody": {
			"content": {
			  "application/json": {
				"schema": {
				  "$ref": "#/components/schemas/AutoRenewalRequest"
				}
			  }
			},
			"required": true
		  },
		  "responses": {
			"200": {
			  "description": "OK",
			  "content": {
				"*/*": {
				  "schema": {
					"$ref": "#/components/schemas/ApiResponseAutoRenewalResponse"
				  }
				}
			  }
			}
		  }
		},
		"delete": {
		  "tags": [
			"plan-commercial-alias-controller"
		  ],
		  "operationId": "disableAutoRenewal",
		  "parameters": [
			{
			  "name": "planCode",
			  "in": "path",
			  "required": true,
			  "schema": {
				"type": "string"
			  }
			},
			{
			  "name": "organizationId",
			  "in": "query",
			  "required": true,
			  "schema": {
				"type": "string",
				"format": "uuid"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "OK",
			  "content": {
				"*/*": {
				  "schema": {
					"$ref": "#/components/schemas/ApiResponseAutoRenewalResponse"
				  }
				}
			  }
			}
		  }
		}
	  },
...
	},
	"components": {
	  "schemas": {
...
		"AutoRenewalRequest": {
		  "type": "object",
		  "properties": {
			"organizationId": {
			  "type": "string",
			  "format": "uuid"
			},
			"addOnCodes": {
			  "type": "array",
			  "items": {
				"type": "string"
			  }
			},
			"billingPeriod": {
			  "type": "string",
			  "enum": [
				"MONTHLY",
				"YEARLY"
			  ]
			},
			"clientId": {
			  "type": "string"
			},
			"provider": {
			  "type": "string",
			  "enum": [
				"MYCOOLPAY",
				"STRIPE"
			  ]
			},
			"method": {
			  "type": "string",
			  "enum": [
				"MOBILE_MONEY",
				"CARD"
			  ]
			},
			"payerReference": {
			  "type": "string"
			}
		  }
		},
		"ApiResponseAutoRenewalResponse": {
		  "type": "object",
		  "properties": {
			"success": {
			  "type": "boolean"
			},
			"data": {
			  "$ref": "#/components/schemas/AutoRenewalResponse"
			},
			"message": {
			  "type": "string"
			},
			"errorCode": {
			  "type": "string"
			},
			"timestamp": {
			  "type": "string",
			  "format": "date-time"
			}
		  }
		},
		"AutoRenewalResponse": {
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string",
			  "format": "uuid"
			},
			"tenantId": {
			  "type": "string",
			  "format": "uuid"
			},
			"organizationId": {
			  "type": "string",
			  "format": "uuid"
			},
			"planCode": {
			  "type": "string"
			},
			"addOnCodes": {
			  "type": "array",
			  "items": {
				"type": "string"
			  }
			},
			"billingPeriod": {
			  "type": "string"
			},
			"clientId": {
			  "type": "string"
			},
			"provider": {
			  "type": "string"
			},
			"method": {
			  "type": "string"
			},
			"payerReference": {
			  "type": "string"
			},
			"active": {
			  "type": "boolean"
			},
			"createdAt": {
			  "type": "string",
			  "format": "date-time"
			},
			"updatedAt": {
			  "type": "string",
			  "format": "date-time"
			}
		  }
		},
...
	  },
	  "securitySchemes": {
		"ClientId": {
		  "type": "apiKey",
		  "description": "Client application identifier for server-to-server access",
		  "name": "X-Client-Id",
		  "in": "header"
		},
		"ApiKey": {
		  "type": "apiKey",
		  "description": "Client application secret for server-to-server access",
		  "name": "X-Api-Key",
		  "in": "header"
		},
		"BearerAuth": {
		  "type": "http",
		  "description": "JWT access token returned by /api/auth/login",
		  "scheme": "bearer",
		  "bearerFormat": "JWT"
		}
	  }
	}
  }